+++ examples/new-project/.loom/shadow/admission-loom.json
@@ -4,6 +4,6 @@
     ".loom/work-items/INIT-0001.md"
   ],
   "source_sha256": {
-    ".loom/work-items/INIT-0001.md": "070e46ddb2e129c7e082398a6123c8f328c7002eca76ed6307b9ef12374d2615"
+    ".loom/work-items/INIT-0001.md": "63069d7da8288e04662ad380f5f4d19b888a3ffc52e56cae6238cabb170cde9d"
   }
 }
