+++ examples/new-project/.loom/work-items/INIT-0001.md
@@ -37,6 +37,7 @@
 - `.gitignore`
 - `.loom/bootstrap/manifest.json`
 - `.loom/bin/failure_envelope.py`
+- `.loom/bin/execution_attempts.py`
 - `.loom/bin/loom_init.py`
 - `.loom/bin/fact_chain_support.py`
 - `.loom/bin/governance_surface.py`
