+++ plugins/loom/.codex-plugin/plugin.json
@@ -23,7 +23,7 @@
     "source_git_sha": "unreleased",
     "source_git_sha_status": "pending_release_commit",
     "plugin_payload_version": "0.28.0",
-    "plugin_payload_hash": "c39fbc0d542896c734f33c132e25d0f8115bcce0f054bc1a89906d459bf3b0c1",
+    "plugin_payload_hash": "87fd6577270f1a427969eb3eb34c4cffcdb8788b85b89b4c42d6b8090fa8c6f6",
     "repository_payload": false
   },
   "keywords": [
