+++ tools/check_npm_package.py
@@ -485,6 +485,7 @@ RUNTIME_COPY_ROOTS = (
 RUNTIME_COPY_FILES = (
     "authority_contract.py",
     "failure_envelope.py",
+    "execution_attempts.py",
     "github_admission.py",
     "github_host.py",
     "loom_init.py",
